dishdash-backend/register/requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
mysql-connector-python==8.2.0
Requests==2.31.0
